Access supabase client through useSupabaseClient
Removes reliance on the outside-react global state of the previous supabase
client instance.

I think this important down the road for SSR with react and supabase, now
the supabase client is entirely managed by NextJS and supabase helpers.

I'm a bit sad that we how have to drill the supabase client through
so many functions to use it. I would love if a reference to it could
live in Zustand stores.